fix(graphviz): Splinesep is C integer division (nodesep/4=4) — 162 byte-exact, corpus 137/138
gv's dot_splines_ initialises .Splinesep = GD_nodesep(g) / 4 where
GD_nodesep is an int (types.h): C integer division, 4 for the default
nodesep 18. The port computed NodeSep / 4.0 = 4.5. (Contrast
maximal_bbox's GD_nodesep(g) / 2. — the trailing dot makes that one
floating division.)

The 0.5pt surfaced through the cluster-wall channel clamp
(maximal_bbox -> cl_bound): round(bb.UR.x + Splinesep) put
162-cluster-style's a0->b0 head-rank box at x=83 vs gv's 82, and the
routed spline pivots exactly at that corridor corner. Found by dumping
gv's routesplines input boxes; the clBound transcription itself was
already faithful.

Promote 162 out of both deferral sets (its fails-when-fixed guard
fired); the single remaining deferral is 03 (intentional, gated vs the
03b newrank oracle). 759 green; gv worktree reverted to pristine.

+++ graphviz/shared/src/main/scala/org/jpablo/graphexplorer/graphviz/layout/Spline.scala
@@ -80,7 +80,12 @@ object Spline:
     // GD_nodesep (attr-driven), the spline channel separation (nodesep/4),
     // and a plain virtual node's half width (1 + nodesep/2). Default 18/4.5/10.
     val NodeSep     = Coord.nodeSepPt(g)
-    val Splinesep   = NodeSep / 4.0
+    // dot_splines_ (dotsplines.c:275): `.Splinesep = GD_nodesep(g) / 4` — with
+    // GD_nodesep an *int* (types.h), this is C INTEGER division: 18/4 = 4, not
+    // 4.5. (Contrast maximal_bbox's `GD_nodesep(g) / 2.` — the trailing dot
+    // makes that one floating division.) The 0.5pt matters: a cluster-wall
+    // clamp `round(bb.urx + Splinesep)` lands at 82 vs 83 (162-cluster-style).
+    val Splinesep   = (NodeSep.toInt / 4).toDouble
     val VirtualHalf = 1.0 + NodeSep / 2.0
     // The Order/XCoord migration to LayoutNode preserves all type-safety at
     // construction. Spline's internals stay String-keyed for its 50+ lookup
